Refactoring: Move some structures used in tinode.conf
