+++ server/main.go
@@ -187,30 +187,6 @@ var globals struct {
 	iceServers []iceServer
 }
 
-// Credential validator config.
-type validatorConfig struct {
-	// TRUE or FALSE to set
-	AddToTags bool `json:"add_to_tags"`
-	//  Authentication level which triggers this validator: "auth", "anon"... or ""
-	Required []string `json:"required"`
-	// Validator params passed to validator unchanged.
-	Config json.RawMessage `json:"config"`
-}
-
-// Large file handler config.
-type mediaConfig struct {
-	// The name of the handler to use for file uploads.
-	UseHandler string `json:"use_handler"`
-	// Maximum allowed size of an uploaded file
-	MaxFileUploadSize int64 `json:"max_size"`
-	// Garbage collection timeout
-	GcPeriod int `json:"gc_period"`
-	// Number of entries to delete in one pass
-	GcBlockSize int `json:"gc_block_size"`
-	// Individual handler config params to pass to handlers unchanged.
-	Handlers map[string]json.RawMessage `json:"handlers"`
-}
-
 // Contentx of the configuration file
 type configType struct {
 	// HTTP(S) address:port to listen on for websocket and long polling clients. Either a
